[Refactor] Isolate LTX operator integration
